fix(client): replace float ltp price cast with Decimal
Why: Fix audit finding [28] in docs/reviews/audit_2026-05-15.md for financial precision
What:
- src/client/upstox_market.py: replace float cast on LTP prices with Decimal and update type hints
- src/client/protocol.py: update get_ltp signature in MarketDataProvider and BrokerClient to return Decimal
- src/client/upstox_live.py: update get_ltp signature in UpstoxLiveClient to return Decimal
- src/client/mock_client.py: update get_ltp, set_price, and internal price map to use Decimal
- tests/unit/test_client.py: import Decimal and update unit tests to assert Decimal return values
- tests/unit/test_mock_client.py: import Decimal and update unit tests to assert Decimal return values
Ref: none

+++ src/client/protocol.py
@@ -33,6 +33,7 @@ composition root) knows which implementation to wire.
 
 from __future__ import annotations
 
+from decimal import Decimal
 from typing import Any, Callable, Protocol, runtime_checkable
 
 from src.models import OptionChain
@@ -66,7 +67,7 @@ class MarketDataProvider(Protocol):
     protocol without any inheritance.
     """
 
-    async def get_ltp(self, instruments: list[str]) -> dict[str, float]: ...
+    async def get_ltp(self, instruments: list[str]) -> dict[str, Decimal]: ...
 
     async def get_option_chain(
         self, instrument: str, expiry: str
@@ -123,7 +124,7 @@ class BrokerClient(Protocol):
 
     # ── MarketDataProvider surface ───────────────────────────────
 
-    async def get_ltp(self, instruments: list[str]) -> dict[str, float]: ...
+    async def get_ltp(self, instruments: list[str]) -> dict[str, Decimal]: ...
 
     async def get_option_chain(
         self, instrument: str, expiry: str
